Move encoding test to its own file
